split library
