+++ ydb/core/tx/columnshard/engines/scheme/versions/ya.make
@@ -1,0 +1,14 @@
+LIBRARY()
+
+SRCS(
+    abstract_scheme.cpp
+    snapshot_scheme.cpp
+    filtered_scheme.cpp
+    versioned_index.cpp
+)
+
+PEERDIR(
+    ydb/core/tx/columnshard/engines/scheme/abstract
+)
+
+END()
